Type flag changed to a fix width (8bit)

+++ src/cl_object_rep.h
@@ -19,10 +19,11 @@
 #ifndef CL_OBJECT_REP_H
 #define CL_OBJECT_REP_H
 
+#include <inttypes.h>
 #include "cl_object.h"
 
 struct cl_object_info_s {
-	size_t _type;
+	uint8_t _type;
 	size_t _ref;
 	cl_object_destructor _dest;
 };
